Add feature to present words in groups
In the file stimuli.js stimuli are added to the experiment.
The stimuli are regular string embedded with special fragments
of text. For example: words marked with a starting "#", are
recorded to the output. Other fragments are not.
This feature adds a string on which sentences are split
into groups. The words in one group are presented together
as group. This is optional, as when no grouping_string
is specified, every word counts as group.